Original XLibs, restored code and newer Vista Engine

+++ XLibs.Net/STLPort/test/eh/random_number.cpp
@@ -1,6 +1,6 @@
 /***********************************************************************************
   random_number.cpp
-
+  
  * Copyright (c) 1997
  * Mark of the Unicorn, Inc.
  *
@@ -23,7 +23,7 @@
 # include <stdlib.h>
 #endif
 
-unsigned random_number( size_t range )
+unsigned random_number( unsigned range )
 {
 #if !defined( __SGI_STL )
   if (range == 0) return 0;
